re-activated the right panel drawer
