+++ src/Routes/index.js
@@ -16,7 +16,6 @@ import {
   useLocalStorage,
   useViewType,
 } from 'hooks';
-// import DashboardDrawer from 'components/Dashboard/Drawer';
 import Header from 'components/Dashboard/Header';
 
 import Body from './Body';
@@ -74,10 +73,6 @@ const Routes = () => {
       {isDataAvailable ? (
         <>
           <UserGuide />
-          {/* 
-          // TODO: TEMPORARY_DEACTIVATED
-          <DashboardDrawer /> 
-          */}
           <Wrapper>
             <LayoutFullHeight>
               <Route path="/:pageName" component={SidebarLeft} />
