+++ src/components/Sidebar/SidebarRight/ErrorLogs/ErrorLogsTable.react.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { useErrorLogs } from 'hooks';
 import { Table } from 'components';
@@ -6,7 +6,11 @@ import { JsonSwitch, Card } from 'components/common';
 import errorLogsTableColumns from './ErrorLogsTableColumns.react';
 
 const ErrorLogsTable = () => {
-  const { dataSource } = useErrorLogs();
+  const { dataSource, setIsCleared } = useErrorLogs();
+
+  useEffect(() => {
+    setIsCleared(true);
+  }, [setIsCleared]);
 
   const expandedRowRender = record => (
     <Card>
