+++ src/components/Sidebar/SidebarRight/SidebarRight.react.js
@@ -1,8 +1,20 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { Layout, Icon, Menu, Badge } from 'antd';
-import { useRightSidebar } from 'hooks';
+import {
+  getBottomActions,
+  topActions,
+} from 'config/schema/rightSidebar.schema';
+import { useDrawer, useErrorLogs } from 'hooks';
+import useStats from 'hooks/useStats';
+import useStorage from 'hooks/useStorage';
+import {
+  getColorStatus,
+  getStorageColorStatus,
+  combineStatus,
+} from 'utils/warningColorStatus';
+import Drawer from './Drawer';
 
 const SiderLight = styled(Layout.Sider)`
   border: none;
@@ -14,43 +26,62 @@ const topMargin = { marginTop: '20%' };
 const noItemSelect = [];
 
 const SidebarRight = ({ isTop, className }) => {
-  const {
-    onSelect,
-    menus: { top, bottom },
-  } = useRightSidebar();
+  const { openDrawer } = useDrawer();
 
-  const menuSelect = useCallback(({ key }) => onSelect(key), [onSelect]);
+  const { totalNewWarnings } = useErrorLogs();
+  const { cpu, memory, gpu } = useStats();
+  const { storage } = useStorage();
+  const { top, bottom } = useMemo(
+    () => ({
+      top: topActions,
+      bottom: getBottomActions({
+        warnings: totalNewWarnings,
+        cpuStatus: getColorStatus(cpu),
+        memoryStatus: combineStatus(
+          getColorStatus(memory),
+          getStorageColorStatus(storage)
+        ),
+        gpuStatus: getColorStatus(gpu),
+      }),
+    }),
+    [cpu, gpu, memory, storage, totalNewWarnings]
+  );
+
+  const menuSelect = useCallback(({ key }) => openDrawer(key), [openDrawer]);
 
   return (
-    <SiderLight
-      className={className}
-      theme="light"
-      collapsedWidth={60}
-      collapsed>
-      <Menu
-        mode="vertical"
-        onSelect={menuSelect}
-        style={topMargin}
-        selectedKeys={noItemSelect}>
-        {(isTop ? top : bottom).map(
-          ({ name, type, component, count, status }) => (
-            <Menu.Item key={name} title={name}>
-              <Badge
-                status={status}
-                count={count}
-                overflowCount={100}
-                offset={[0, 11]}>
-                <Icon
-                  type={type}
-                  component={component}
-                  style={centerIconStyle}
-                />
-              </Badge>
-            </Menu.Item>
-          )
-        )}
-      </Menu>
-    </SiderLight>
+    <>
+      <Drawer />
+      <SiderLight
+        className={className}
+        theme="light"
+        collapsedWidth={60}
+        collapsed>
+        <Menu
+          mode="vertical"
+          onSelect={menuSelect}
+          style={topMargin}
+          selectedKeys={noItemSelect}>
+          {(isTop ? top : bottom).map(
+            ({ name, type, component, count, status }) => (
+              <Menu.Item key={name} title={name}>
+                <Badge
+                  status={status}
+                  count={count}
+                  overflowCount={100}
+                  offset={[0, 11]}>
+                  <Icon
+                    type={type}
+                    component={component}
+                    style={centerIconStyle}
+                  />
+                </Badge>
+              </Menu.Item>
+            )
+          )}
+        </Menu>
+      </SiderLight>
+    </>
   );
 };
