+++ src/hooks/useDrawer.js
@@ -1,14 +1,31 @@
-import { STATE_SOURCES } from 'const';
-import { useActions } from 'hooks';
-import { useSelector } from 'react-redux';
+import { useCallback, useState } from 'react';
 import { createStore } from 'reusable';
 
 const useDrawer = () => {
-  const { isVisible, content } = useSelector(
-    state => state[STATE_SOURCES.DRAWER]
+  /* 
+    there's are separate values for the selected panel and the visibility
+    the reason is to keep the items on the page while the drawer is animating out
+    do not use the selected panel as a visibility check!
+   */
+  const [{ selectedPanel, isVisible }, setSelectedPanel] = useState({
+    selectedPanel: null,
+    isVisible: false,
+  });
+
+  const openDrawer = useCallback(
+    selection =>
+      setSelectedPanel({
+        selectedPanel: selection,
+        isVisible: true,
+      }),
+    []
   );
-  const { drawerOpen } = useActions();
-  return { isVisible, content, drawerOpen };
+  const closeDrawer = useCallback(
+    () => setSelectedPanel(state => ({ ...state, isVisible: false })),
+    [setSelectedPanel]
+  );
+
+  return { isVisible, selectedPanel, openDrawer, closeDrawer };
 };
 
 export default createStore(useDrawer);
